# Conflicts:
#	app/src/main/java/a75f/io/renatus/profiles/system/advancedahu/AdvancedHybridAhuViewModel.kt

+++ app/src/main/java/a75f/io/renatus/composables/CmComposables.kt
@@ -264,7 +264,7 @@ fun AnalogOutConfig(
     unit: String = "",
     isEnabled: Boolean,
     onAssociationChanged: (Option) -> Unit = {},
-    testVal: Int,
+    testVal: Double,
     onTestSignalSelected: (Double) -> Unit = {}
 ) {
     Row(
